Refactor collection creation dialogs and logic for dynamic roles
