+++ app/src/main/java/com/nuvio/tv/ui/screens/cast/CastDetailScreen.kt
@@ -157,7 +157,7 @@ fun CastDetailScreen(
 
         if (showCollectionDialog && personDetail != null) {
             if (showNameInputDialog) {
-                CollectionNameInputDialog(
+                com.nuvio.tv.ui.components.CollectionNameInputDialog(
                     initialValue = "",
                     onConfirm = { customName, viewMode ->
                         viewModel.createNewCollection(customName, personDetail, viewMode)
@@ -165,7 +165,7 @@ fun CastDetailScreen(
                     onDismiss = { showNameInputDialog = false }
                 )
             } else {
-                CastCollectionPickerDialog(
+                com.nuvio.tv.ui.components.CollectionPickerDialog(
                     collections = existingCollections,
                     onCreateNewClick = { showNameInputDialog = true },
                     onCollectionClick = { collectionId ->
@@ -772,220 +772,5 @@ private fun formatDateForDisplay(date: String?): String? {
         SimpleDateFormat(android.text.format.DateFormat.getBestDateTimePattern(locale, "dMMMy"), locale).format(parsed)
     } catch (_: Exception) {
         null
-    }
-}
-
-@OptIn(ExperimentalTvMaterial3Api::class)
-@Composable
-private fun CastCollectionPickerDialog(
-    collections: List<com.nuvio.tv.domain.model.Collection>,
-    onCreateNewClick: () -> Unit,
-    onCollectionClick: (String) -> Unit,
-    onDismiss: () -> Unit
-) {
-    val primaryFocusRequester = remember { FocusRequester() }
-
-    LaunchedEffect(Unit) {
-        primaryFocusRequester.requestFocus()
-    }
-
-    NuvioDialog(
-        onDismiss = onDismiss,
-        title = stringResource(R.string.cast_detail_dialog_choose_collection),
-        subtitle = stringResource(R.string.cast_detail_dialog_choose_collection_subtitle)
-    ) {
-        Button(
-            onClick = onCreateNewClick,
-            modifier = Modifier
-                .fillMaxWidth()
-                .focusRequester(primaryFocusRequester),
-            colors = ButtonDefaults.colors(
-                containerColor = NuvioTheme.colors.BackgroundCard,
-                contentColor = NuvioTheme.colors.TextPrimary
-            )
-        ) {
-            Text(stringResource(R.string.cast_detail_dialog_create_new))
-        }
-
-        if (collections.isNotEmpty()) {
-            val listState = rememberLazyListState()
-            Box(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .heightIn(max = 240.dp)
-            ) {
-                LazyColumn(
-                    state = listState,
-                    modifier = Modifier.fillMaxWidth(),
-                    verticalArrangement = Arrangement.spacedBy(NuvioTheme.spacing.xs)
-                ) {
-                    items(collections) { collection ->
-                        Button(
-                            onClick = { onCollectionClick(collection.id) },
-                            modifier = Modifier.fillMaxWidth(),
-                            colors = ButtonDefaults.colors(
-                                containerColor = NuvioTheme.colors.BackgroundCard,
-                                contentColor = NuvioTheme.colors.TextPrimary
-                            )
-                        ) {
-                            Text(collection.title)
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-@OptIn(ExperimentalTvMaterial3Api::class)
-@Composable
-private fun CollectionNameInputDialog(
-    initialValue: String,
-    onConfirm: (String, FolderViewMode) -> Unit,
-    onDismiss: () -> Unit
-) {
-    var name by remember { mutableStateOf(initialValue) }
-    var viewMode by remember { mutableStateOf(FolderViewMode.TABBED_GRID) }
-    val focusRequester = remember { FocusRequester() }
-    val keyboardController = LocalSoftwareKeyboardController.current
-    var isEditing by remember { mutableStateOf(false) }
-
-    fun isSelectKey(keyCode: Int): Boolean {
-        return keyCode == AndroidKeyEvent.KEYCODE_DPAD_CENTER ||
-            keyCode == AndroidKeyEvent.KEYCODE_ENTER ||
-            keyCode == AndroidKeyEvent.KEYCODE_NUMPAD_ENTER
-    }
-
-    LaunchedEffect(Unit) {
-        focusRequester.requestFocus()
-    }
-
-    NuvioDialog(
-        onDismiss = onDismiss,
-        title = "Create New Collection",
-        subtitle = "Enter name & select format",
-        width = 460.dp
-    ) {
-        androidx.compose.material3.OutlinedTextField(
-            value = name,
-            onValueChange = { name = it },
-            modifier = Modifier
-                .fillMaxWidth()
-                .focusRequester(focusRequester)
-                .onFocusChanged {
-                    if (!it.isFocused) {
-                        isEditing = false
-                    }
-                }
-                .onPreviewKeyEvent { event ->
-                    val native = event.nativeKeyEvent
-                    if (native.action == AndroidKeyEvent.ACTION_DOWN && isSelectKey(native.keyCode)) {
-                        isEditing = true
-                        keyboardController?.show()
-                    }
-                    false
-                },
-            readOnly = !isEditing,
-            singleLine = true,
-            maxLines = 1,
-            keyboardOptions = KeyboardOptions(imeAction = ImeAction.Done),
-            keyboardActions = KeyboardActions(
-                onDone = {
-                    isEditing = false
-                    keyboardController?.hide()
-                }
-            ),
-            label = { androidx.compose.material3.Text("Collection Name") },
-            colors = androidx.compose.material3.OutlinedTextFieldDefaults.colors(
-                focusedTextColor = NuvioTheme.colors.TextPrimary,
-                unfocusedTextColor = NuvioTheme.colors.TextPrimary,
-                focusedContainerColor = NuvioTheme.colors.BackgroundCard,
-                unfocusedContainerColor = NuvioTheme.colors.BackgroundCard,
-                focusedBorderColor = NuvioTheme.colors.FocusRing,
-                unfocusedBorderColor = NuvioTheme.colors.Border,
-                focusedLabelColor = NuvioTheme.colors.TextSecondary,
-                unfocusedLabelColor = NuvioTheme.colors.TextTertiary,
-                cursorColor = NuvioTheme.colors.FocusRing
-            )
-        )
-
-        Spacer(modifier = Modifier.height(NuvioTheme.spacing.md))
-
-        androidx.compose.material3.Text(
-            text = "Collection Format",
-            style = MaterialTheme.typography.labelMedium,
-            color = NuvioTheme.colors.TextSecondary,
-            modifier = Modifier.padding(bottom = NuvioTheme.spacing.xs)
-        )
-
-        Row(
-            modifier = Modifier.fillMaxWidth().padding(bottom = NuvioTheme.spacing.md),
-            horizontalArrangement = Arrangement.spacedBy(NuvioTheme.spacing.md)
-        ) {
-            val gridShape = RoundedCornerShape(NuvioTheme.radii.sm)
-            val rowsShape = RoundedCornerShape(NuvioTheme.radii.sm)
-
-            Card(
-                onClick = { viewMode = FolderViewMode.TABBED_GRID },
-                modifier = Modifier.weight(1f).height(44.dp),
-                shape = CardDefaults.shape(shape = gridShape),
-                colors = CardDefaults.colors(
-                    containerColor = if (viewMode == FolderViewMode.TABBED_GRID) NuvioTheme.colors.FocusRing.copy(alpha = 0.15f) else NuvioTheme.colors.BackgroundCard,
-                    focusedContainerColor = NuvioTheme.colors.BackgroundCard
-                ),
-                border = CardDefaults.border(
-                    border = Border(border = BorderStroke(1.dp, if (viewMode == FolderViewMode.TABBED_GRID) NuvioTheme.colors.FocusRing else NuvioTheme.colors.Border), shape = gridShape),
-                    focusedBorder = Border(border = BorderStroke(2.dp, NuvioTheme.colors.FocusRing), shape = gridShape)
-                )
-            ) {
-                Box(modifier = Modifier.fillMaxSize(), contentAlignment = Alignment.Center) {
-                    Text("Tabbed Grid", style = MaterialTheme.typography.labelMedium)
-                }
-            }
-
-            Card(
-                onClick = { viewMode = FolderViewMode.ROWS },
-                modifier = Modifier.weight(1f).height(44.dp),
-                shape = CardDefaults.shape(shape = rowsShape),
-                colors = CardDefaults.colors(
-                    containerColor = if (viewMode == FolderViewMode.ROWS) NuvioTheme.colors.FocusRing.copy(alpha = 0.15f) else NuvioTheme.colors.BackgroundCard,
-                    focusedContainerColor = NuvioTheme.colors.BackgroundCard
-                ),
-                border = CardDefaults.border(
-                    border = Border(border = BorderStroke(1.dp, if (viewMode == FolderViewMode.ROWS) NuvioTheme.colors.FocusRing else NuvioTheme.colors.Border), shape = rowsShape),
-                    focusedBorder = Border(border = BorderStroke(2.dp, NuvioTheme.colors.FocusRing), shape = rowsShape)
-                )
-            ) {
-                Box(modifier = Modifier.fillMaxSize(), contentAlignment = Alignment.Center) {
-                    Text("Rows", style = MaterialTheme.typography.labelMedium)
-                }
-            }
-        }
-
-        Button(
-            onClick = {
-                if (name.isNotBlank()) {
-                    onConfirm(name, viewMode)
-                }
-            },
-            modifier = Modifier.fillMaxWidth(),
-            colors = ButtonDefaults.colors(
-                containerColor = NuvioTheme.colors.BackgroundCard,
-                contentColor = NuvioTheme.colors.TextPrimary
-            )
-        ) {
-            Text("Create")
-        }
-
-        Button(
-            onClick = onDismiss,
-            modifier = Modifier.fillMaxWidth(),
-            colors = ButtonDefaults.colors(
-                containerColor = NuvioTheme.colors.BackgroundCard,
-                contentColor = NuvioTheme.colors.TextPrimary
-            )
-        ) {
-            Text("Cancel")
-        }
     }
 }
